Move node out of its subdirectory.
