+++ src/nodes/energy.cpp
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (c) 2026 Team Dissolve and contributors
 
-#include "nodes/energy/energy.h"
+#include "nodes/energy.h"
 #include "base/sysFunc.h"
 #include "kernels/producer.h"
 #include "main/dissolve.h"
