+++ src/nodes/energy.h
