+++ tests/ff/energy-water-SPCFw.cpp
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (c) 2026 Team Dissolve and contributors
 
+#include "nodes/energy.h"
 #include "nodes/gr/gr.h"
 #include "tests/graphData.h"
 #include "tests/testData.h"
